backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
torch
diffusers
transformers
accelerate
pillow
numpy
opencv-python
protobuf
sentencepiece
huggingface_hub
# Trigger deployment
